requirements.txt
pytesseract
Pillow
python-dotenv
opencv-python
numpy
easyocr
pdf2image
pandas
openpyxl
